requirements.txt
langchain_openai
langchain_core
langchain
python-dotenv
streamlit
langchain_community
uvicorn
sse_starlette
huggingface_hub 
tiktoken
sentence_transformers==2.2.2
pinecone-client
flask
markdown
ctransformers==0.2.5
-e .